+++ lib/diamond/src/c_api/mod.rs
@@ -1,0 +1,4 @@
+mod builder;
+mod collision;
+mod geometry;
+mod mesh;
