+++ lib/diamond/src/mesh/mod.rs
@@ -1,0 +1,6 @@
+pub mod builder;
+pub mod collision;
+
+pub struct Mesh {
+	pub collision: collision::Mesh,
+}
